shift action to src, split for easier testing
