+++ packages/action/package.json
@@ -17,6 +17,7 @@
   },
   "devDependencies": {
     "@vercel/ncc": "^0.27.0",
+    "fixturez": "^1.1.0",
     "tslib": "^2.1.0",
     "typescript": "^4.2.3"
   },
