+++ packages/covector/package.json
@@ -35,6 +35,7 @@
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-typescript": "^8.2.0",
     "@types/inquirer": "^7.3.1",
+    "fixturez": "^1.1.0",
     "rollup": "^2.42.4",
     "tslib": "^2.1.0",
     "typescript": "^4.2.3"
